+++ lib/openapi_first/response_validator.rb
@@ -10,57 +10,13 @@ module OpenapiFirst
     def initialize(spec)
       @spec = spec
       @router = Router.new(->(_env) {}, spec: spec, raise_error: true)
+      @response_validation = ResponseValidation.new(->(response) { response.to_a })
     end
 
     def validate(request, response)
-      errors = validation_errors(request, response)
-      Validation.new(errors || [])
-    rescue OpenapiFirst::ResponseCodeNotFoundError, OpenapiFirst::NotFoundError => e
-      Validation.new([e.message])
-    end
-
-    def validate_operation(request, response)
-      errors = validation_errors(request, response)
-      Validation.new(errors || [])
-    rescue OpenapiFirst::ResponseCodeNotFoundError, OpenapiFirst::NotFoundError => e
-      Validation.new([e.message])
-    end
-
-    private
-
-    def validation_errors(request, response)
-      content = response_for(request, response)&.fetch('content', nil)
-      return unless content
-
-      content_type = content[response.content_type]
-      return ["Content type not found: '#{response.content_type}'"] unless content_type
-
-      response_schema = content_type['schema']
-      return unless response_schema
-
-      response_data = MultiJson.load(response.body)
-      validate_json_schema(response_schema, response_data)
-    end
-
-    def validate_json_schema(schema, data)
-      JSONSchemer.schema(schema).validate(data).to_a.map do |error|
-        format_error(error)
-      end
-    end
-
-    def format_error(error)
-      ValidationFormat.error_details(error)
-                      .merge!(
-                        data_pointer: error['data_pointer'],
-                        schema_pointer: error['schema_pointer']
-                      )
-    end
-
-    def response_for(request, response)
       env = request.env.dup
       @router.call(env)
-      operation = env[OPERATION]
-      operation&.response_for(response.status)
+      @response_validation.validate(response, env[OPERATION])
     end
   end
 end