+++ spec/response_validator_spec.rb
@@ -22,38 +22,31 @@ RSpec.describe OpenapiFirst::ResponseValidator do
   let(:headers) { { Rack::CONTENT_TYPE => 'application/json' } }
 
   describe 'valid response' do
-    it 'returns no errors' do
+    it 'raises nothing' do
       response_body = json_dump([
                                   { id: 42, name: 'hans' },
                                   { id: 2, name: 'Voldemort' }
                                 ])
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be false
-      expect(result.errors).to be_empty
+      subject.validate(request, response)
     end
 
     it 'falls back to the default' do
       response_body = JSON.dump(code: 422, message: 'Not good!')
       response = Rack::MockResponse.new(422, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be false
-      expect(result.errors).to be_empty
+      subject.validate(request, response)
     end
 
     it 'returns no errors on additional, not required properties' do
       response_body = json_dump([{ id: 42, name: 'hans', something: 'else' }])
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors).to be_empty
+      subject.validate(request, response)
     end
 
     it 'returns no errors if OAS file has no content' do
       expect_any_instance_of(OpenapiFirst::Operation).to receive(:response_for) { {} }
       response = Rack::MockResponse.new(200, headers, 'body')
-      result = subject.validate(request, response)
-      expect(result.errors?).to be false
-      expect(result.errors).to be_empty
+      subject.validate(request, response)
     end
 
     it 'returns no errors if OAS file has no response_for schema specified' do
@@ -61,9 +54,7 @@ RSpec.describe OpenapiFirst::ResponseValidator do
       expect_any_instance_of(OpenapiFirst::Operation)
         .to receive(:response_for) { { 'content' => empty_content } }
       response = Rack::MockResponse.new(200, headers, 'body')
-      result = subject.validate(request, response)
-      expect(result.errors?).to be false
-      expect(result.errors).to be_empty
+      subject.validate(request, response)
     end
   end
 
@@ -75,9 +66,9 @@ RSpec.describe OpenapiFirst::ResponseValidator do
       end
       response_body = json_dump([{ id: 'string', name: 'hans' }])
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be true
-      expect(result.errors.first).to eq "Could not find definition for PATCH '/pets' in API description ./spec/data/petstore.yaml" # rubocop:disable Layout/LineLength
+      expect do
+        subject.validate(request, response)
+      end.to raise_error OpenapiFirst::NotFoundError
     end
 
     it 'fails on unknown status' do
@@ -85,37 +76,34 @@ RSpec.describe OpenapiFirst::ResponseValidator do
       request = Rack::Request.new(env)
       response_body = json_dump([{ id: 2, name: 'Voldemort' }])
       response = Rack::MockResponse.new(201, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be true
-      operation_name = 'GET /pets/{petId} (showPetById)'
-      expect(result.errors.first).to eq "Response status code or default not found: 201 for '#{operation_name}'"
+      expect do
+        subject.validate(request, response)
+      end.to raise_error OpenapiFirst::ResponseInvalid
     end
 
     it 'fails on wrong content type' do
       response_body = json_dump([{ id: 2, name: 'Voldemort' }])
       headers = { Rack::CONTENT_TYPE => 'application/xml' }
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be true
-      expect(result.errors.first).to eq(
-        "Content type not found: 'application/xml'"
-      )
+      expect do
+        subject.validate(request, response)
+      end.to raise_error OpenapiFirst::ResponseInvalid
     end
 
     it 'returns errors on missing property' do
       response_body = json_dump([{ id: 42 }, { id: 2, name: 'Voldemort' }])
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors?).to be true
-      error = result.errors.first
-      expect(error[:title]).to eq('is missing required properties: name')
+      expect do
+        subject.validate(request, response)
+      end.to raise_error OpenapiFirst::ResponseInvalid
     end
 
     it 'returns errors on wrong property type' do
       response_body = json_dump([{ id: 'string', name: 'hans' }])
       response = Rack::MockResponse.new(200, headers, response_body)
-      result = subject.validate(request, response)
-      expect(result.errors.first[:title]).to eq('should be a integer')
+      expect do
+        subject.validate(request, response)
+      end.to raise_error OpenapiFirst::ResponseInvalid
     end
   end
 end