+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Change `ResponseValidator` to raise an exception if it found a problem
 - Params have symbolized keys now
 - Remove `not_found` option from Router. Return 405 if HTTP verb is not allowed (via Hanami::Router)
 - Add `raise_error` option to OpenapiFirst.app (false by default)